[level 0] Title: 간단한 식 계산하기, Time: 1.13 ms, Memory: 47.7 MB -BaekjoonHub

+++ C#/프로그래머스/0/181865. 간단한 식 계산하기/간단한 식 계산하기.cs
@@ -6,8 +6,8 @@ public class Solution
     {
         long answer = 0;
         string[] op = binomial.Split(' ');
-        long a = int.Parse(op[0]);
-        long b = int.Parse(op[2]);
+        long a = long.Parse(op[0]);
+        long b = long.Parse(op[2]);
         if(op[1] == "+") answer = a + b;
         else if(op[1] == "-") answer = a - b;
         else if(op[1] == "*") answer = a * b;
